backend/requirements.txt
fastapi[standard]
pydantic
pydantic-settings
instructor
openai
python-multipart
pytesseract
Pillow
apscheduler
emails
pytest
requests
pypdf
